Split logger JS API spec tests into browser vs non-browser (skipping a few in browser for now)
